+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Minor
 
-- IconButton: Add new `disabled` prop and stylings to `IconButton` component (#517)
+- IconButton: Add new `disabled` prop and stylings to `IconButton` component (#521)
 
 ### Patch